+++ SkillModifier_Regs.lua
@@ -6,9 +6,6 @@ SkillModifier.register_modifier("echo_item", 3200, function(skill, data, item_id
     local stop_frame = 0
     SkillModifier.add_on_activate_callback(data, function(skill_)
         local current_frame = gm.variable_global_get("_current_frame")
-        if current_frame <= stop_frame then
-            return false
-        end
         local base = math.max(skill_.cooldown_base, 10) -- use_next_frame_
         if current_frame - last_frame >= base and stack < 4 then
             stack = stack + 1
@@ -21,7 +18,7 @@ SkillModifier.register_modifier("echo_item", 3200, function(skill, data, item_id
                 gm.item_take(skill_.parent, item_id, stack, 0)
             end
             stack = 0
-            stop_frame = current_frame + (base + 30) * 6
+            skill_.use_next_frame = current_frame + (base + 30) * 6
         end, math.floor(base + 30))
     end)
 end, function(skill, data)
@@ -61,8 +58,11 @@ SkillModifier.register_modifier("life_burn", 250, function(skill, data)
                 if skill_.stock < skill_.max_stock then
                     Utils.sync_call("gml_Script_actor_skill_add_stock", "host", skill_.parent, skill_.slot_index)
                     gm.call("gml_Script_actor_skill_add_stock", skill_.parent, skill_.slot_index)
-                    local num = Utils.get_handy_drone_type(skill_.skill_id) ~= nil and 25 or skill_.cooldown / 60 * 5
-                    Utils.set_and_sync_inst_from_table(skill_.parent,{hp = skill_.parent.hp - num})
+                    local num = Utils.get_handy_drone_type(skill_.skill_id) ~= nil and 25 or skill_.cooldown_base / 60 *
+                                    5
+                    Utils.set_and_sync_inst_from_table(skill_.parent, {
+                        hp = skill_.parent.hp - num
+                    })
                 end
             end
         end
@@ -70,7 +70,13 @@ SkillModifier.register_modifier("life_burn", 250, function(skill, data)
 end, function(skill, data)
     SkillModifier.remove_on_can_activate_callback(data)
 end)
-
+SkillModifier.register_modifier("after_image", 3200, function(skill, data, item_id)
+    SkillModifier.add_on_activate_callback(data, function(skill_)
+        gm.apply_buff(skill_.parent, 44.0, skill_.cooldown_base)
+    end)
+end, function(skill, data)
+    SkillModifier.remove_on_activate_callback(data)
+end)
 local function register_buffer(attr, fn)
     SkillModifier.register_modifier("flux_" .. attr, 125, function(skill, data, value)
         SkillModifier.change_attr(skill, attr, data, value)